♻️ Moving Client.on('ready') to its own event file
